#329 move all simplify tests to new file
