+++ tests/unit/test_expression_tree/test_symbol.py
@@ -8,7 +8,6 @@ import pybamm
 import unittest
 import numpy as np
 import os
-import math
 
 
 class TestSymbol(unittest.TestCase):
@@ -38,195 +37,6 @@ class TestSymbol(unittest.TestCase):
 
         # check that the *actual* children are updated
         check_are_equal(super(pybamm.Symbol, symp).children, (symc1, symc2, symc3))
-
-    def test_symbol_simplify(self):
-        a = pybamm.Scalar(0)
-        b = pybamm.Scalar(1)
-        d = pybamm.Scalar(-1)
-        e = pybamm.Scalar(2)
-
-        # negate
-        self.assertIsInstance((-a).simplify(), pybamm.Scalar)
-        self.assertEqual((-a).simplify().evaluate(), 0)
-        self.assertIsInstance((-b).simplify(), pybamm.Scalar)
-        self.assertEqual((-b).simplify().evaluate(), -1)
-
-        # absolute value
-        self.assertIsInstance((abs(a)).simplify(), pybamm.Scalar)
-        self.assertEqual((abs(a)).simplify().evaluate(), 0)
-        self.assertIsInstance((abs(d)).simplify(), pybamm.Scalar)
-        self.assertEqual((abs(d)).simplify().evaluate(), 1)
-
-        # function
-        def sin(x):
-            return math.sin(x)
-
-        f = pybamm.Function(sin, b)
-        self.assertIsInstance((f).simplify(), pybamm.Scalar)
-        self.assertEqual((f).simplify().evaluate(), math.sin(1))
-
-        # Gradient
-        self.assertIsInstance((pybamm.grad(a)).simplify(), pybamm.Scalar)
-        self.assertEqual((pybamm.grad(a)).simplify().evaluate(), 0)
-        v = pybamm.Variable("v")
-        self.assertIsInstance((pybamm.grad(v)).simplify(), pybamm.Gradient)
-
-        # Divergence
-        self.assertIsInstance((pybamm.div(a)).simplify(), pybamm.Scalar)
-        self.assertEqual((pybamm.div(a)).simplify().evaluate(), 0)
-        self.assertIsInstance((pybamm.div(v)).simplify(), pybamm.Divergence)
-
-        # Integral
-        self.assertIsInstance(
-            (pybamm.Integral(a, pybamm.t)).simplify(), pybamm.Integral
-        )
-
-        # BoundaryValue
-        v_neg = pybamm.Variable("v", domain=["negative electrode"])
-        self.assertIsInstance((pybamm.surf(v_neg)).simplify(), pybamm.BoundaryValue)
-
-        # addition
-        self.assertIsInstance((a + b).simplify(), pybamm.Scalar)
-        self.assertEqual((a + b).simplify().evaluate(), 1)
-        self.assertIsInstance((b + b).simplify(), pybamm.Scalar)
-        self.assertEqual((b + b).simplify().evaluate(), 2)
-        self.assertIsInstance((b + a).simplify(), pybamm.Scalar)
-        self.assertEqual((b + a).simplify().evaluate(), 1)
-
-        # subtraction
-        self.assertIsInstance((a - b).simplify(), pybamm.Scalar)
-        self.assertEqual((a - b).simplify().evaluate(), -1)
-        self.assertIsInstance((b - b).simplify(), pybamm.Scalar)
-        self.assertEqual((b - b).simplify().evaluate(), 0)
-        self.assertIsInstance((b - a).simplify(), pybamm.Scalar)
-        self.assertEqual((b - a).simplify().evaluate(), 1)
-
-        # multiplication
-        self.assertIsInstance((a * b).simplify(), pybamm.Scalar)
-        self.assertEqual((a * b).simplify().evaluate(), 0)
-        self.assertIsInstance((b * a).simplify(), pybamm.Scalar)
-        self.assertEqual((b * a).simplify().evaluate(), 0)
-        self.assertIsInstance((b * b).simplify(), pybamm.Scalar)
-        self.assertEqual((b * b).simplify().evaluate(), 1)
-        self.assertIsInstance((a * a).simplify(), pybamm.Scalar)
-        self.assertEqual((a * a).simplify().evaluate(), 0)
-
-        # test when other node is a parameter
-        c = pybamm.Parameter("c")
-        self.assertIsInstance((a + c).simplify(), pybamm.Parameter)
-        self.assertIsInstance((c + a).simplify(), pybamm.Parameter)
-        self.assertIsInstance((c + b).simplify(), pybamm.Addition)
-        self.assertIsInstance((b + c).simplify(), pybamm.Addition)
-        self.assertIsInstance((a * c).simplify(), pybamm.Scalar)
-        self.assertEqual((a * c).simplify().evaluate(), 0)
-        self.assertIsInstance((c * a).simplify(), pybamm.Scalar)
-        self.assertEqual((c * a).simplify().evaluate(), 0)
-        self.assertIsInstance((b * c).simplify(), pybamm.Parameter)
-        self.assertIsInstance((e * c).simplify(), pybamm.Multiplication)
-
-        expr = (e * (e * c)).simplify()
-        self.assertIsInstance(expr, pybamm.Multiplication)
-        self.assertIsInstance(expr.children[0], pybamm.Scalar)
-        self.assertIsInstance(expr.children[1], pybamm.Parameter)
-
-        expr = (e / (e * c)).simplify()
-        self.assertIsInstance(expr, pybamm.Division)
-        self.assertIsInstance(expr.children[0], pybamm.Scalar)
-        self.assertEqual(expr.children[0].evaluate(), 1.0)
-        self.assertIsInstance(expr.children[1], pybamm.Parameter)
-
-        expr = (e * (e / c)).simplify()
-        self.assertIsInstance(expr, pybamm.Division)
-        self.assertIsInstance(expr.children[0], pybamm.Scalar)
-        self.assertEqual(expr.children[0].evaluate(), 4.0)
-        self.assertIsInstance(expr.children[1], pybamm.Parameter)
-
-        expr = (e * (c / e)).simplify()
-        self.assertIsInstance(expr, pybamm.Multiplication)
-        self.assertIsInstance(expr.children[0], pybamm.Scalar)
-        self.assertEqual(expr.children[0].evaluate(), 1.0)
-        self.assertIsInstance(expr.children[1], pybamm.Parameter)
-
-        expr = ((e * c) * (c / e)).simplify()
-        self.assertIsInstance(expr, pybamm.Multiplication)
-        self.assertIsInstance(expr.children[0], pybamm.Scalar)
-        self.assertEqual(expr.children[0].evaluate(), 1.0)
-        self.assertIsInstance(expr.children[1], pybamm.Multiplication)
-        self.assertIsInstance(expr.children[1].children[0], pybamm.Parameter)
-        self.assertIsInstance(expr.children[1].children[1], pybamm.Parameter)
-
-        expr = (e + (e + c)).simplify()
-        self.assertIsInstance(expr, pybamm.Addition)
-        self.assertIsInstance(expr.children[0], pybamm.Scalar)
-        self.assertEqual(expr.children[0].evaluate(), 4.0)
-        self.assertIsInstance(expr.children[1], pybamm.Parameter)
-
-        expr = (e + (e - c)).simplify()
-        self.assertIsInstance(expr, pybamm.Subtraction)
-        self.assertIsInstance(expr.children[0], pybamm.Scalar)
-        self.assertEqual(expr.children[0].evaluate(), 4.0)
-        self.assertIsInstance(expr.children[1], pybamm.Parameter)
-
-        expr = ((2 + c) + (c + 2)).simplify()
-        self.assertIsInstance(expr, pybamm.Addition)
-        self.assertIsInstance(expr.children[0], pybamm.Scalar)
-        self.assertEqual(expr.children[0].evaluate(), 4.0)
-        self.assertIsInstance(expr.children[1], pybamm.Addition)
-        self.assertIsInstance(expr.children[1].children[0], pybamm.Parameter)
-        self.assertIsInstance(expr.children[1].children[1], pybamm.Parameter)
-
-        expr = ((-1 + c) - (c + 1) + (c - 1)).simplify()
-        self.assertIsInstance(expr, pybamm.Addition)
-        self.assertIsInstance(expr.children[0], pybamm.Scalar)
-        self.assertEqual(expr.children[0].evaluate(), -3.0)
-
-        # check these don't simplify
-        self.assertIsInstance((c * e).simplify(), pybamm.Multiplication)
-        self.assertIsInstance((e / c).simplify(), pybamm.Division)
-        self.assertIsInstance((c).simplify(), pybamm.Parameter)
-        c1 = pybamm.Parameter("c1")
-        self.assertIsInstance((c1 * c).simplify(), pybamm.Multiplication)
-
-        # should simplify division to multiply
-        self.assertIsInstance((c / e).simplify(), pybamm.Multiplication)
-
-        self.assertIsInstance((c / b).simplify(), pybamm.Parameter)
-        self.assertIsInstance((c * b).simplify(), pybamm.Parameter)
-
-        # negation with parameter
-        self.assertIsInstance((-c).simplify(), pybamm.Negate)
-
-        self.assertIsInstance((a + b + a).simplify(), pybamm.Scalar)
-        self.assertEqual((a + b + a).simplify().evaluate(), 1)
-        self.assertIsInstance((b + a + a).simplify(), pybamm.Scalar)
-        self.assertEqual((b + a + a).simplify().evaluate(), 1)
-        self.assertIsInstance((a * b * b).simplify(), pybamm.Scalar)
-        self.assertEqual((a * b * b).simplify().evaluate(), 0)
-        self.assertIsInstance((b * a * b).simplify(), pybamm.Scalar)
-        self.assertEqual((b * a * b).simplify().evaluate(), 0)
-
-        # power simplification
-        self.assertIsInstance((c ** a).simplify(), pybamm.Scalar)
-        self.assertEqual((c ** a).simplify().evaluate(), 1)
-        self.assertIsInstance((a ** c).simplify(), pybamm.Scalar)
-        self.assertEqual((a ** c).simplify().evaluate(), 0)
-        d = pybamm.Scalar(2)
-        self.assertIsInstance((c ** d).simplify(), pybamm.Power)
-
-        # division
-        self.assertIsInstance((a / b).simplify(), pybamm.Scalar)
-        self.assertEqual((a / b).simplify().evaluate(), 0)
-        self.assertIsInstance((b / a).simplify(), pybamm.Scalar)
-        self.assertEqual((b / a).simplify().evaluate(), np.inf)
-        self.assertIsInstance((a / a).simplify(), pybamm.Scalar)
-        self.assertTrue(np.isnan((a / a).simplify().evaluate()))
-        self.assertIsInstance((b / b).simplify(), pybamm.Scalar)
-        self.assertEqual((b / b).simplify().evaluate(), 1)
-
-        # not implemented for Symbol
-        sym = pybamm.Symbol("sym")
-        with self.assertRaises(NotImplementedError):
-            sym.simplify()
 
     def test_symbol_domains(self):
         a = pybamm.Symbol("a", domain=pybamm.KNOWN_DOMAINS[0])
